aws-smithy-http-client: clarify pool state ownership
Extract the origin demand schedule behind its own invariant-owning module and place state definitions before the implementations that mutate them. Document the HTTP/1 ownership paths and tighten the surrounding module narrative so reviewers can follow each terminal transition.
